Fixed the linter errors
